+++ src/common.js
@@ -161,6 +161,12 @@ const PREDEFINE_SHORTCUTS = {
     bucket: 'eventcertainties',
     params: mergeObjects(DEFAULT_PARAMS),
   },
+  eventAction: {
+    shortcut: 'eventAction',
+    wellknown: 'predefine',
+    bucket: 'eventactions',
+    params: mergeObjects(DEFAULT_PARAMS),
+  },
   eventFunction: {
     shortcut: 'eventFunction',
     wellknown: 'predefine',
@@ -171,6 +177,18 @@ const PREDEFINE_SHORTCUTS = {
     shortcut: 'eventGroup',
     wellknown: 'predefine',
     bucket: 'eventgroups',
+    params: mergeObjects(DEFAULT_PARAMS),
+  },
+  eventIndicator: {
+    shortcut: 'eventIndicator',
+    wellknown: 'predefine',
+    bucket: 'eventindicators',
+    params: mergeObjects(DEFAULT_PARAMS),
+  },
+  eventQuestion: {
+    shortcut: 'eventQuestion',
+    wellknown: 'predefine',
+    bucket: 'eventquestions',
     params: mergeObjects(DEFAULT_PARAMS),
   },
   eventSeverity: {
@@ -225,6 +243,12 @@ const PREDEFINE_SHORTCUTS = {
     shortcut: 'partyRole',
     wellknown: 'predefine',
     bucket: 'partyroles',
+    params: mergeObjects(DEFAULT_PARAMS),
+  },
+  unit: {
+    shortcut: 'unit',
+    wellknown: 'predefine',
+    bucket: 'units',
     params: mergeObjects(DEFAULT_PARAMS),
   },
 };